+++ duologsync/producer/telephony_producer.py
@@ -1,9 +1,14 @@
 """
 Definition of the TelephonyProducer class
 """
+import datetime
+import functools
+import math
 
 from duologsync.config import Config
 from duologsync.producer.producer import Producer
+from duologsync.util import normalize_params, run_in_executor
+
 
 class TelephonyProducer(Producer):
     """
@@ -11,7 +16,51 @@ class TelephonyProducer(Producer):
     and placement into a queue of Telephony logs
     """
 
-    def __init__(self, api_call, log_queue, child_account_id=None, url_path=None):
-        super().__init__(api_call, log_queue, Config.TELEPHONY,
-                         account_id=child_account_id,
-                         url_path=url_path)
+    def __init__(self, api_call, log_queue, url_path=None):
+        super().__init__(
+            api_call,
+            log_queue,
+            Config.TELEPHONY,
+            url_path=url_path,
+        )
+        self.mintime = None
+
+        # Default to the generated mintime if no checkpoint is found
+        if isinstance(self.log_offset, int):
+            self.mintime = self.log_offset
+            self.log_offset = None
+
+        # If we have a string based offset, it's pulled from the checkpoint file
+        if isinstance(self.log_offset, str):
+            (previous_time, _) = self.log_offset.split(",")
+            self.mintime = int(previous_time)
+            self.log_offset = None
+
+    async def call_log_api(self):
+        """
+        Make a call to the Telephony Log API endpoint and return the result
+
+        @return the result of a call to the Telephony Log API endpoint
+        """
+        today = datetime.datetime.now(tz=datetime.timezone.utc)
+        maxtime = math.floor(today.timestamp() * 1000) - 120
+
+        parameters = normalize_params(
+            {
+                "mintime": f"{self.mintime}",
+                "maxtime": f"{maxtime}",
+                "limit": "1000",
+                "sort": "ts:asc",
+            }
+        )
+
+        if self.log_offset is not None:
+            parameters["next_offset"] = [f"{self.log_offset}"]
+
+        api_result = await run_in_executor(
+            functools.partial(
+                self.api_call, method="GET", path=self.url_path, params=parameters
+            )
+        )
+
+        return api_result